timers: move priority_queue to internal/timers/
