+++ lib/internal/timers.js
@@ -102,7 +102,7 @@ const {
 const { validateNumber } = require('internal/validators');
 
 const L = require('internal/linkedlist');
-const PriorityQueue = require('internal/priority_queue');
+const PriorityQueue = require('internal/timers/priority_queue');
 
 const { inspect } = require('internal/util/inspect');
 let debuglog;
