+++ lib/internal/timers/priority_queue.js
