+++ node.gyp
@@ -157,7 +157,6 @@
       'lib/internal/options.js',
       'lib/internal/policy/manifest.js',
       'lib/internal/policy/sri.js',
-      'lib/internal/priority_queue.js',
       'lib/internal/process/esm_loader.js',
       'lib/internal/process/execution.js',
       'lib/internal/process/main_thread_only.js',
@@ -179,6 +178,7 @@
       'lib/internal/socket_list.js',
       'lib/internal/test/binding.js',
       'lib/internal/timers.js',
+      'lib/internal/timers/priority_queue.js',
       'lib/internal/tls.js',
       'lib/internal/trace_events_async_hooks.js',
       'lib/internal/tty.js',
