+++ test/parallel/test-bootstrap-modules.js
@@ -46,7 +46,6 @@ const expectedModules = new Set([
   'NativeModule internal/modules/cjs/helpers',
   'NativeModule internal/modules/cjs/loader',
   'NativeModule internal/options',
-  'NativeModule internal/priority_queue',
   'NativeModule internal/process/execution',
   'NativeModule internal/process/per_thread',
   'NativeModule internal/process/promises',
@@ -54,6 +53,7 @@ const expectedModules = new Set([
   'NativeModule internal/process/warning',
   'NativeModule internal/querystring',
   'NativeModule internal/timers',
+  'NativeModule internal/timers/priority_queue',
   'NativeModule internal/url',
   'NativeModule internal/util',
   'NativeModule internal/util/debuglog',
